RUNCONQUER-FINAL/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0
scikit-learn==1.5.0
numpy==2.0.0
joblib==1.4.0
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==24.1.0
